Move Content job classes to `content` namespace
