+++ app/jobs/content/import_content_item_job.rb
@@ -1,0 +1,8 @@
+module Content
+  class ImportContentItemJob < ApplicationJob
+    def perform(*args)
+      content_id = args[0]
+      Importers::SingleContentItem.run(content_id)
+    end
+  end
+end
