--- app/jobs/import_content_item_job.rb
@@ -1,6 +1,0 @@
-class ImportContentItemJob < ApplicationJob
-  def perform(*args)
-    content_id = args[0]
-    Importers::SingleContentItem.run(content_id)
-  end
-end
