--- spec/jobs/import_content_item_job_spec.rb
@@ -1,6 +1,0 @@
-RSpec.describe ImportContentItemJob, type: :job do
-  it "calls the single content item importer" do
-    expect(Importers::SingleContentItem).to receive(:run).with("id-123")
-    subject.perform("id-123")
-  end
-end
